+++ code/src/vscode-dts/vscode.proposed.terminalCompletionProvider.d.ts
@@ -19,11 +19,12 @@ declare module 'vscode' {
 		provideTerminalCompletions(terminal: Terminal, context: TerminalCompletionContext, token: CancellationToken): ProviderResult<T[] | TerminalCompletionList<T>>;
 	}
 
+
 	export interface TerminalCompletionItem {
 		/**
 		 * The label of the completion.
 		 */
-		label: string;
+		label: string | CompletionItemLabel;
 
 		/**
 		 * The index of the start of the range to replace.
@@ -59,10 +60,12 @@ declare module 'vscode' {
 	export enum TerminalCompletionItemKind {
 		File = 0,
 		Folder = 1,
-		Flag = 2,
-		Method = 3,
+		Method = 2,
+		Alias = 3,
 		Argument = 4,
-		Alias = 5,
+		Option = 5,
+		OptionValue = 6,
+		Flag = 7,
 	}
 
 	export interface TerminalCompletionContext {
@@ -125,6 +128,10 @@ declare module 'vscode' {
 		 * Show folders as completion items.
 		 */
 		foldersRequested?: boolean;
+		/**
+		 * File extensions to filter by.
+		 */
+		fileExtensions?: string[];
 		/**
 		 * If no cwd is provided, no resources will be shown as completions.
 		 */